fix: Forward errors to main isolate in renderHtml
Errors when calling renderHtml were silently failing and resulting in timeouts because the exception wasn't automatically passed to the spawning isolate. Add an separate `errorPort` to suppport sending the error to the calling isolate which will properly throw it back the the handler.

This probably also needs to be done for `renderData`.

refs: #129